+++ assistants/core/src/rules/layer-styles-no-dirty/__tests__/index.ts
@@ -4,8 +4,7 @@ describe('layer-styles-no-dirty', () => {
   test('outputs no violations for layers that do not differ from their shared style', async (): Promise<
     void
   > => {
-    expect.assertions(2)
-    const { violations, errors } = await testRule(
+    const { violations, ruleErrors } = await testRule(
       __dirname,
       './all_good.sketch',
       'layer-styles-no-dirty',
@@ -14,13 +13,13 @@ describe('layer-styles-no-dirty', () => {
       },
     )
     expect(violations).toHaveLength(0)
-    expect(errors).toHaveLength(0)
+    expect(ruleErrors).toHaveLength(0)
   })
+
   test('finds violations for layer styles that differ from their shared styles', async (): Promise<
     void
   > => {
-    expect.assertions(2)
-    const { violations, errors } = await testRule(
+    const { violations, ruleErrors } = await testRule(
       __dirname,
       './rectangles.sketch',
       'layer-styles-no-dirty',
@@ -29,12 +28,11 @@ describe('layer-styles-no-dirty', () => {
       },
     )
     expect(violations).toHaveLength(1)
-    expect(errors).toHaveLength(0)
+    expect(ruleErrors).toHaveLength(0)
   })
 
   test('outputs no violations for artboards and pages', async (): Promise<void> => {
-    expect.assertions(2)
-    const { violations, errors } = await testRule(
+    const { violations, ruleErrors } = await testRule(
       __dirname,
       './artboards-and-pages.sketch',
       'layer-styles-no-dirty',
@@ -43,12 +41,11 @@ describe('layer-styles-no-dirty', () => {
       },
     )
     expect(violations).toHaveLength(0)
-    expect(errors).toHaveLength(0)
+    expect(ruleErrors).toHaveLength(0)
   })
 
   test('outputs no violations for text layers', async (): Promise<void> => {
-    expect.assertions(2)
-    const { violations, errors } = await testRule(
+    const { violations, ruleErrors } = await testRule(
       __dirname,
       './text-layers.sketch',
       'layer-styles-no-dirty',
@@ -57,14 +54,13 @@ describe('layer-styles-no-dirty', () => {
       },
     )
     expect(violations).toHaveLength(0)
-    expect(errors).toHaveLength(0)
+    expect(ruleErrors).toHaveLength(0)
   })
 
   test('finds violations for groups that differ from their shared style', async (): Promise<
     void
   > => {
-    expect.assertions(2)
-    const { violations, errors } = await testRule(
+    const { violations, ruleErrors } = await testRule(
       __dirname,
       './groups_bad.sketch',
       'layer-styles-no-dirty',
@@ -73,6 +69,6 @@ describe('layer-styles-no-dirty', () => {
       },
     )
     expect(violations).toHaveLength(1)
-    expect(errors).toHaveLength(0)
+    expect(ruleErrors).toHaveLength(0)
   })
 })