+++ assistants/core/src/rules/text-styles-prefer-library/__tests__/index.ts
@@ -2,8 +2,7 @@ import { testRule } from '../../../test-helpers'
 
 describe('text-styles-prefer-library', () => {
   test('finds no violations if only library styles are used', async (): Promise<void> => {
-    expect.assertions(2)
-    const { violations, errors } = await testRule(
+    const { violations, ruleErrors } = await testRule(
       __dirname,
       './all-good.sketch',
       'text-styles-prefer-library',
@@ -13,14 +12,13 @@ describe('text-styles-prefer-library', () => {
       },
     )
     expect(violations).toHaveLength(0)
-    expect(errors).toHaveLength(0)
+    expect(ruleErrors).toHaveLength(0)
   })
 
   test('reports violations when a text style differs from its library style', async (): Promise<
     void
   > => {
-    expect.assertions(2)
-    const { violations, errors } = await testRule(
+    const { violations, ruleErrors } = await testRule(
       __dirname,
       './text-differs.sketch',
       'text-styles-prefer-library',
@@ -30,12 +28,11 @@ describe('text-styles-prefer-library', () => {
       },
     )
     expect(violations).toHaveLength(1)
-    expect(errors).toHaveLength(0)
+    expect(ruleErrors).toHaveLength(0)
   })
 
   test('reports violations when a text does not have a library style', async (): Promise<void> => {
-    expect.assertions(2)
-    const { violations, errors } = await testRule(
+    const { violations, ruleErrors } = await testRule(
       __dirname,
       './text-no-library.sketch',
       'text-styles-prefer-library',
@@ -45,12 +42,11 @@ describe('text-styles-prefer-library', () => {
       },
     )
     expect(violations).toHaveLength(2)
-    expect(errors).toHaveLength(0)
+    expect(ruleErrors).toHaveLength(0)
   })
 
   test('reports violations if an unauthorized library is used', async (): Promise<void> => {
-    expect.assertions(2)
-    const { violations, errors } = await testRule(
+    const { violations, ruleErrors } = await testRule(
       __dirname,
       './all-good.sketch',
       'text-styles-prefer-library',
@@ -60,6 +56,6 @@ describe('text-styles-prefer-library', () => {
       },
     )
     expect(violations).toHaveLength(2) // one violation per unauthorized library use
-    expect(errors).toHaveLength(0)
+    expect(ruleErrors).toHaveLength(0)
   })
 })