+++ packages/utils/CHANGELOG.md
@@ -1,6 +1,25 @@
 # @sketch-hq/sketch-assistant-utils
 
+## 5.0.0-next.14
+
+### Patch Changes
+
+- a2a4585: Do not throw an error when processing a Sketch file and encountering an unknown object
+  `_class` value.
+
+## 5.0.0-next.11
+
+### Major Changes
+
+- 6733a97: Numerous refactors around Sketch file object iteration.
+
+### Patch Changes
+
+- Updated dependencies [6733a97]
+  - @sketch-hq/sketch-assistant-types@5.0.0-next.11
+
 ## 4.0.1-next.8
+
 ### Patch Changes
 
 - bc13811: Initial release in monorepo.